+++ jdbc/src/main/java/com/interface21/jdbc/callback/PreparedStatementCallBack.java
@@ -1,4 +1,4 @@
-package com.interface21.jdbc;
+package com.interface21.jdbc.callback;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
